backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
moviepy==1.0.3
gTTS==2.3.2
google-generativeai==0.3.0
python-dotenv==1.0.0